requirements.txt
flask==3.0.0
requests==2.31.0
python-dateutil==2.8.2
xmltodict==0.13.0
